+++ src/onegov.core/onegov/core/orm/abstract/__init__.py
@@ -1,0 +1,5 @@
+from onegov.core.orm.abstract.adjacency_list import (
+    AdjacencyList, AdjacencyListCollection
+)
+
+__all__ = ['AdjacencyList', 'AdjacencyListCollection']